requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow
fastapi
uvicorn
python-multipart

